moved BaseCommandInterface.php in Contract directory
